+++ MagRetentionReload.Net/Properties/AssemblyInfo.cs
@@ -1,0 +1,11 @@
+using System.Reflection;
+using System.Runtime.InteropServices;
+using Tosox.MagRetentionReload;
+
+[assembly: AssemblyTitle(ModInfo.Name + " - Fika packets")]
+[assembly: AssemblyCompany(ModInfo.Source)]
+[assembly: AssemblyProduct(ModInfo.Name)]
+[assembly: AssemblyCopyright("Copyright © " + ModInfo.Author + " 2026")]
+[assembly: AssemblyVersion(ModInfo.PacketsVersion)]
+[assembly: AssemblyFileVersion(ModInfo.PacketsVersion)]
+[assembly: ComVisible(false)]